requirements.txt
Jinja2==3.1.4
SQLAlchemy==2.0.34
email-validator==2.2.0
fastapi==0.115.0
pydantic==2.9.2
python-multipart==0.0.9
requests==2.32.3
uvicorn[standard]==0.30.6
python-dotenv
